+++ Filter.h
@@ -7,18 +7,18 @@
 class Filter {
 public:
     Filter();
-    size_t getWidth();
+    double getWidth();
     std::string& getBorder();
     std::string& getType();
     double getValue(double x);
-    void setWidth(size_t w);
+    void setWidth(double w);
     void setBorder(std::string& b);
     void setType(std::string& newType);
 
 private:
     std::string type;
     std::string border;
-    size_t width;
+    double width;
 
     double sinc(double x);
 };
